+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Introduce a method named `asString()` to the `Vector*` classes to return a string representation of the vectors that
   will not be optimized out if it is used, regardless of the compile time settings.
 * Create a minimal `Matrix4d` matrix.
+* Add a `toArray()` method to the `Vector*` classes to support extracting the vector components as an array of values.
 
 ### [v0.07](https://github.com/realityforge/vecmath/tree/v0.07) (2020-10-18)
 [Full Changelog](https://github.com/realityforge/vecmath/compare/v0.06...v0.07)